requirements.txt
flask==3.1.0
anthropic==0.52.0
requests==2.32.3
python-dotenv==1.0.1
pillow==11.1.0
